fix: use gogoproto instead of gogo/protobuf in Filter
The Filter function was using github.com/gogo/protobuf/proto which is
incompatible with cosmos-sdk v0.50 types that use github.com/cosmos/gogoproto/proto.

This caused proto.MessageName() to return empty strings for gogoproto types,
resulting in all filters matching all events (since "" == ""), which caused
events to be routed to the wrong handlers.

The panic observed was:
  job panicked: interface conversion: interface {} is *types.KeygenStarted, not *types.SigningStarted

This occurred because KeygenStarted events were incorrectly passing the
SigningStarted filter and being sent to the signing job handler.

Fixes the Filter function by using the correct gogoproto package that is
compatible with cosmos-sdk v0.50.